+++ internal/queue.go
@@ -20,7 +20,7 @@ type Queue interface {
 	CreateQueue(queue_name string) error
 	DeleteQueue(queue_name string) error
 	Enqueue(queue_name string, item interface{}) error
-	EnqueueBatch(queue_name string, items []interface{}) (int, error)
+	EnqueueBatch(queue_name string, items []interface{}) (int64, error)
 	Dequeue(queue_name string, group_name string, consumer_id string) (QueueMessage, error)
 	Ack(queue_name string, group_name string, messageID int64, receipt string) error
 	Nack(queue_name string, group_name string, messageID int64, receipt string) error
